requirements.txt
fastapi
uvicorn
faiss-cpu>=1.9.0
sentence-transformers
transformers
beautifulsoup4
requests
numpy
